clean up remove promo screen
