+++ app/lib/screens/remove_card/remove_card_error.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class RemoveCardError extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: ListView(
+        children: [
+          Text('error'),
+        ],
+      ),
+    );
+  }
+}
