+++ app/lib/screens/remove_card/removing_card.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class RemovingCard extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
