+++ app/lib/screens/remove_promo/remove_promo_error.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class RemovePromoError extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text(
+          'Error :(',
+          style: TextStyle(color: Colors.cyan),
+        ),
+      ),
+    );
+  }
+  
+}
