+++ app/lib/screens/remove_promo/removing_promo.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class RemovingPromo extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+  
+}
